Move OssPackage and OssPackageResults types to config.go for better organization
